Apply rustfmt and split completion tests
